% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coastline}
\alias{coastline}
\title{Coastline shape of the UK, used to differentiate land and sea in maps}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 1 columns.
}
\usage{
coastline
}
\description{
Coastline shape of the UK, used to differentiate land and sea in maps
}
\keyword{datasets}
