% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_box.R
\name{create_box}
\alias{create_box}
\title{Create rectangular boundary box}
\usage{
create_box(location, y_dist = 2500, x_dist = y_dist * 2/3)
}
\arguments{
\item{location}{A vector consisting of easting and northing,
e.g. \code{c(349950, 673605)}}

\item{y_dist}{Distance (in metres) north and south of the location
to be displayed. Map shows \code{y_dist} metres north, and
\code{y_dist} metres south.}

\item{x_dist}{Distance (in metres) east and west of the location
to be displayed. Map shows \code{y_dist} metres north, and
\code{y_dist} metres south.}
}
\value{
An st polygon with four points.
}
\description{
Create rectangular boundary centred on a set location
}
\examples{

create_box(location = c(349950, 673605), x_dist = 3000, y_dist = 1000)
}
