% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_circle.R
\name{create_circle}
\alias{create_circle}
\title{Create circular boundary box}
\usage{
create_circle(location, radius = 1000)
}
\arguments{
\item{location}{A vector consisting of easting and northing,
e.g. \code{c(349950, 673605)}}

\item{radius}{Radius (in metres) around the location to be displayed.}
}
\value{
A circular st polygon.
}
\description{
Create circular boundary centred on a set location
}
\examples{

create_circle(location = c(349950, 673605), radius = 1500)
}
