% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map.R
\name{create_map}
\alias{create_map}
\title{Create aesthetic map of area around a postcode}
\usage{
create_map(
  postcode,
  shape = "rectangle",
  shapefiles_folder = "/Users/Shared/data/shapefiles/local_map",
  y_dist = 2500,
  x_dist = y_dist * 2/3,
  radius = 1500,
  water_colour = "#C0C0C0",
  land_colour = "#002240",
  road_colour = "#C0C0C0"
)
}
\arguments{
\item{postcode}{A UK postcode, e.g. "EH3 7HG"}

\item{shape}{\code{"circle"} or \code{"rectangle"} to be drawn.
\code{"circle"} uses \code{\link{create_circle()}}
\code{"rectangle"} uses \code{\link{create_box()}}}

\item{shapefiles_folder}{Folder containing OS local map shapefile}

\item{y_dist}{If shape = \code{"rectangle"}, distance (in metres) north and
south of the location to be displayed. Map shows \code{y_dist} metres north,
and \code{y_dist} metres south.}

\item{x_dist}{If shape = \code{"rectangle"}, distance (in metres) east and
west of the location to be displayed. Map shows \code{y_dist} metres east,
and \code{y_dist} metres west.}

\item{radius}{If shape = \code{"circle"}, radius (in metres) around the
location to be displayed.}

\item{water_colour}{Hex code for colour to display water areas}

\item{land_colour}{Hex code for colour to display land}

\item{road_colour}{Hex code for colour to display roads}
}
\value{
A ggplot object
}
\description{
Create aesthetic map of area around a postcode
}
