% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_postcode_centre.R
\name{find_postcode_centre}
\alias{find_postcode_centre}
\title{Find the centroid of a postcode}
\usage{
find_postcode_centre(postcode)
}
\arguments{
\item{postcode}{A UK postcode, e.g. EH41 4BN}
}
\value{
A vector containing easting and northing, e.g.
\code{c(352066, 673922)}
}
\description{
Uses \link{https://postcodes.io} to find centre coordinates
of a postcode area
}
\examples{
find_postcode_centre("EH41 4BN")
}
