% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_square.R
\name{find_square}
\alias{find_square}
\title{Find the OS square for a location}
\usage{
find_square(location)
}
\arguments{
\item{location}{A vector consisting of easting and northing,
e.g. \code{c(349950, 673605)}}
}
\value{

}
\description{
Find the OS 100km grid square containig the location based on easting
and northing.
}
\details{
For information on the 100km grid, see
\url{https://www.ordnancesurvey.co.uk/documents/resources/guide-to-nationalgrid.pdf}
}
\examples{

find_square(c(349950, 673605))
}
