% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_road_data.R
\name{read_road_data}
\alias{read_road_data}
\title{Read and combine OS road data for selected 100km square(s)}
\usage{
read_road_data(folder, squares)
}
\arguments{
\item{folder}{Folder containing OS maps}

\item{squares}{Character vector containing names of OS 100km squares needed.}
}
\value{
ST object of roads
}
\description{
Reads road data from OS local map, combines multiple squares (if present)
into one, and sets the sizes for roads to draw in ggplot
}
\examples{
\dontrun{
read_road_data("/documents", c("NT", "NU"))
}
}
