% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/required_squares.R
\name{required_squares}
\alias{required_squares}
\title{Find all OS squares contained within a boundary}
\usage{
required_squares(boundary)
}
\arguments{
\item{boundary}{An \code{st} object of a single row,
containing a single polygon.}
}
\value{
A vector containing the names of OS squares required to draw
the area.
}
\description{
Uses \code{\link{find_square()}} to find the OS 100km squares required to
draw a selected area.
}
\details{
Creates \code{bbox} around the input \code{boundary} polygon, and looks
for the boxes containg the four corners.
}
